# Conflicts:
#	src/main/java/spring_boot/project_swp/service/impl/PaymentServiceImpl.java
#	src/main/java/spring_boot/project_swp/vnpay/VNPayReturnController.java

+++ src/main/java/spring_boot/project_swp/entity/RentalDiscounts.java
@@ -33,7 +33,7 @@ public class RentalDiscounts {
     Discount discount;
 
     @Column(name = "AppliedAmount", precision = 10, scale = 2)
-    BigDecimal appliedAmount;
+    int appliedAmount;
 
     @Column(name = "AppliedAt")
     LocalDateTime appliedAt;
